moved templates -> dynamic-list-template
